refactor : member 파일명 변경
