+++ backend/src/main/java/org/scoula/member/dto/MemberFcmTokenDTO.java
@@ -8,7 +8,7 @@ import javax.validation.constraints.Size;
 
 @Getter
 @Setter
-public class FcmTokenRequestDTO {
+public class MemberFcmTokenDTO {
     @NotBlank(message = "FCM 토큰을 입력해주세요.")
     @Size(max = 255, message = "FCM 토큰은 255자 이내여야 합니다.")
     private String fcmToken;
