+++ backend/src/main/java/org/scoula/member/dto/MemberLoginRequestDTO.java
@@ -10,7 +10,7 @@ import javax.validation.constraints.NotBlank;
 @Getter
 @Setter
 @NoArgsConstructor
-public class LoginRequestDTO {
+public class MemberLoginRequestDTO {
     @NotBlank(message = "이메일을 입력해주세요.")
     @Email(message = "올바른 이메일 형식이 아닙니다.")
     private String email;
