+++ backend/src/main/java/org/scoula/member/dto/MemberNicknameCheckDTO.java
@@ -10,7 +10,7 @@ import javax.validation.constraints.Size;
 @Getter
 @Setter
 @NoArgsConstructor
-public class NicknameCheckDTO {
+public class MemberNicknameCheckDTO {
     @NotBlank(message = "닉네임을 입력해주세요.")
     @Size(min = 2, max = 100, message = "닉네임은 2-100자 사이여야 합니다.")
     private String nickname;
