+++ backend/src/main/java/org/scoula/member/dto/MemberResponseDTO.java
@@ -1,4 +1,4 @@
-package org.scoula.member.domain;
+package org.scoula.member.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,13 +9,11 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 @NoArgsConstructor
-public class UserVO {
+public class MemberResponseDTO {
+    // 민감 정보(password, fcmToken, phoneNumber) 제외
     private int userId;
     private String email;
-    private String password;
     private String nickname;
-    private LocalDateTime createdAt;
-    private String fcmToken;
     private String name;
-    private String phoneNumber;
+    private LocalDateTime createdAt;
 }
