+++ backend/src/main/java/org/scoula/member/dto/MemberUpdateDTO.java
@@ -12,7 +12,7 @@ import javax.validation.constraints.Size;
 @Getter
 @Setter
 @NoArgsConstructor
-public class UserUpdateDTO {
+public class MemberUpdateDTO {
     // 수정할 nickname
     @NotBlank(message = "닉네임을 입력해주세요.")
     @Size(min = 2, max = 100, message = "닉네임은 2-100자 사이여야 합니다.")
